Issue #1404: Notification service classes moved to acl package
