+++ api/src/main/java/com/epam/pipeline/controller/notification/NotificationController.java
@@ -20,7 +20,7 @@ import com.epam.pipeline.controller.AbstractRestController;
 import com.epam.pipeline.controller.Result;
 import com.epam.pipeline.controller.vo.notification.NotificationMessageVO;
 import com.epam.pipeline.entity.notification.NotificationMessage;
-import com.epam.pipeline.manager.notification.NotificationApiService;
+import com.epam.pipeline.acl.notification.NotificationApiService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
