+++ api/src/main/java/com/epam/pipeline/controller/notification/SystemNotificationController.java
@@ -25,7 +25,7 @@ import com.epam.pipeline.controller.vo.SystemNotificationFilterVO;
 import com.epam.pipeline.entity.notification.SystemNotification;
 import com.epam.pipeline.entity.notification.SystemNotificationConfirmation;
 import com.epam.pipeline.entity.notification.SystemNotificationConfirmationRequest;
-import com.epam.pipeline.manager.notification.SystemNotificationApiService;
+import com.epam.pipeline.acl.notification.SystemNotificationApiService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
