+++ api/src/test/java/com/epam/pipeline/test/web/ControllerTestBeans.java
@@ -48,6 +48,14 @@ import com.epam.pipeline.manager.notification.NotificationApiService;
 import com.epam.pipeline.manager.notification.NotificationSettingsApiService;
 import com.epam.pipeline.manager.notification.NotificationTemplateApiService;
 import com.epam.pipeline.manager.notification.SystemNotificationApiService;
+import com.epam.pipeline.manager.issue.IssueApiService;
+import com.epam.pipeline.manager.metadata.CategoricalAttributeApiService;
+import com.epam.pipeline.manager.metadata.MetadataApiService;
+import com.epam.pipeline.manager.metadata.MetadataEntityApiService;
+import com.epam.pipeline.acl.notification.NotificationApiService;
+import com.epam.pipeline.acl.notification.NotificationSettingsApiService;
+import com.epam.pipeline.acl.notification.NotificationTemplateApiService;
+import com.epam.pipeline.acl.notification.SystemNotificationApiService;
 import com.epam.pipeline.manager.pipeline.FolderApiService;
 import com.epam.pipeline.manager.pipeline.PipelineConfigApiService;
 import com.epam.pipeline.manager.pipeline.ToolManager;
